Move subscription content to a component.
