+++ static/js/src/advantage/react/components/Subscriptions/SubscriptionDetails/DetailsContent/DetailsContent.test.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { shallow } from "enzyme";
+
+import DetailsContent from "./DetailsContent";
+
+describe("DetailsContent", () => {
+  it("renders", () => {
+    const wrapper = shallow(<DetailsContent />);
+    expect(wrapper.find("DetailsTabs").exists()).toBe(true);
+  });
+});
