+++ static/js/src/advantage/react/components/Subscriptions/SubscriptionDetails/DetailsContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DetailsContent";
